+++ server/templates/base.html
@@ -41,7 +41,9 @@ content - main page content
     window.dataLayer = window.dataLayer || [];
     function gtag() { dataLayer.push(arguments); }
     gtag('js', new Date());
-    gtag('config', '{{ GA_ACCOUNT }}', { client_storage: 'none', anonymize_ip: true, place_category: '{{ place_category }}' });
+    gtag('config', '{{ GA_ACCOUNT }}', {
+      client_storage: 'none', anonymize_ip: true, {{ ("place_category: '{}'"|safe).format(place_category) if place_category is defined else ""}}
+    });
 
   </script>
   {% endif %}
